Fixed up the calendar.
Made the calendar better overall. Current date is now has a bold outline, and will be center at the start of the app. Add button to add meals, meals can now be deleted, and renamed. Added button to show more days.
Fixed my spelling of calendar because it is not spelled calendard